+++ src/main/java/com/woowacourse/coordinate/domain/Line.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate;
+package com.woowacourse.coordinate.domain;
 
 import java.util.HashSet;
 import java.util.List;
