+++ src/main/java/com/woowacourse/coordinate/domain/PointDuplicateException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate;
+package com.woowacourse.coordinate.domain;
 
 public class PointDuplicateException extends RuntimeException {
 }
