+++ src/main/java/com/woowacourse/coordinate/domain/Shape.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate;
+package com.woowacourse.coordinate.domain;
 
 public interface Shape {
 	double calculateArea();
