+++ src/main/java/com/woowacourse/coordinate/domain/Square.java
@@ -1,4 +1,4 @@
-package com.woowacourse.coordinate;
+package com.woowacourse.coordinate.domain;
 
 import java.util.List;
 import java.util.Objects;
