+++ src/main/java/com/woowacourse/coordinate/domain/Triangle.java
@@ -1,5 +1,4 @@
-package com.woowacourse.coordinate;
-
+package com.woowacourse.coordinate.domain;
 
 import java.util.List;
 import java.util.Objects;
