+++ src/main/java/com/woowacourse/coordinate/view/InputView.java
@@ -1,9 +1,8 @@
-package com.woowacourse.coordinate;
+package com.woowacourse.coordinate.view;
 
 import java.util.Scanner;
 
 public class InputView {
-
 	public static String inputCoordinate() {
 		Scanner scanner = new Scanner(System.in);
 		System.out.println("좌표를 입력하세요.");
