+++ src/test/java/com/woowacourse/coordinate/PointTest.java
@@ -1,7 +1,6 @@
 package com.woowacourse.coordinate;
 
-import java.util.Arrays;
-
+import com.woowacourse.coordinate.domain.Point;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
