+++ src/test/java/com/woowacourse/coordinate/SquareTest.java
@@ -2,6 +2,8 @@ package com.woowacourse.coordinate;
 
 import java.util.Arrays;
 
+import com.woowacourse.coordinate.domain.Point;
+import com.woowacourse.coordinate.domain.Square;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
